+++ diagrams-demo-gallery/.storybook/theme.js
@@ -1,0 +1,7 @@
+import { create } from '@storybook/theming';
+
+export default create({
+	base: 'dark',
+	brandTitle: 'STORM React Diagrams',
+	brandUrl: 'https://github.com/projectstorm/react-diagrams'
+});