+++ packages/react-canvas-core/src/Toolkit.ts
@@ -18,15 +18,12 @@ export class Toolkit {
 		});
 	}
 
-	/**
-	 * Finds the closest element as a polyfill
-	 */
 	public static closest(element: Element, selector: string) {
 		if (!Element.prototype.closest) {
 			Element.prototype.closest = function(s) {
 				var el = this;
-		  
-				do {
+
+        do {
 					if (Element.prototype.matches.call(el, s)) return el;
 					el = el.parentElement || el.parentNode;
 				} while (el !== null && el.nodeType === 1);