+++ packages/react-canvas-core/src/entities/selection/SelectionBoxWidget.tsx
@@ -16,6 +16,9 @@ namespace S {
 export class SelectionBoxWidget extends React.Component<SelectionBoxWidgetProps> {
 	render() {
 		const { rect } = this.props;
+
+		if (!rect) return null;
+
 		return (
 			<S.Container
 				style={{