+++ packages/react-canvas-core/src/states/SelectionBoxState.ts
@@ -1,7 +1,7 @@
 import { AbstractDisplacementState, AbstractDisplacementStateEvent } from '../core-state/AbstractDisplacementState';
 import { State } from '../core-state/State';
 import { SelectionLayerModel } from '../entities/selection/SelectionLayerModel';
-import { Rectangle } from '@projectstorm/geometry';
+import { Point, Rectangle } from '@projectstorm/geometry';
 import { ModelGeometryInterface } from '../core/ModelGeometryInterface';
 
 export class SelectionBoxState extends AbstractDisplacementState {
@@ -26,7 +26,14 @@ export class SelectionBoxState extends AbstractDisplacementState {
 	}
 
 	getBoxDimensions(event: AbstractDisplacementStateEvent): ClientRect {
-		const rel = this.engine.getRelativePoint(event.event.clientX, event.event.clientY);
+		let rel: Point;
+
+		if ('touches' in event.event) {
+			const touch = event.event.touches[0];
+			rel = this.engine.getRelativePoint(touch.clientX, touch.clientY);
+		} else {
+			rel = this.engine.getRelativePoint(event.event.clientX, event.event.clientY);
+		}
 
 		return {
 			left: rel.x > this.initialXRelative ? this.initialXRelative : rel.x,
@@ -54,8 +61,8 @@ export class SelectionBoxState extends AbstractDisplacementState {
 		const rect = new Rectangle(relative, Math.abs(event.virtualDisplacementX), Math.abs(event.virtualDisplacementY));
 
 		for (let model of this.engine.getModel().getSelectionEntities()) {
-			if (((model as unknown) as ModelGeometryInterface).getBoundingBox) {
-				const bounds = ((model as unknown) as ModelGeometryInterface).getBoundingBox();
+			if ((model as unknown as ModelGeometryInterface).getBoundingBox) {
+				const bounds = (model as unknown as ModelGeometryInterface).getBoundingBox();
 				if (rect.containsPoint(bounds.getTopLeft()) && rect.containsPoint(bounds.getBottomRight())) {
 					model.setSelected(true);
 				} else {